refactor: ыефкеstart to use hooks results
