+++ src/pages/details/Details.tsx
@@ -1,12 +1,10 @@
-import { useNavigate, useParams } from 'react-router-dom';
-import React, { useEffect, useRef } from 'react';
+import { useNavigate } from 'react-router-dom';
+import React, { useRef } from 'react';
 import styles from './Details.module.css';
 import { ROUTES } from '../../routs/routes';
-import { useAppDispatch, useAppSelector } from '../../hooks/redux';
 import { ProductDetails } from '../../components/productDetails/ProductDetails';
 import { createSearchParams } from '../../utils/createSearchParams';
-import { useGetProductByIdQuery } from '../../services/api';
-import { productActions } from '../../reducers/ProductSlice';
+import { useAppSelector } from '../../hooks/redux';
 
 enum TEXTS {
   BUTTON_CLOSE = 'Close',
@@ -15,17 +13,9 @@ enum TEXTS {
 export default function Details() {
   const navigate = useNavigate();
   const overlayRef = useRef<HTMLDivElement>(null);
-  const { productId } = useParams();
   const appSearchParams = useAppSelector(
     (state) => state.appSearchParamsReducer
   );
-
-  const { data: product } = useGetProductByIdQuery(productId || '1');
-  const dispatch = useAppDispatch();
-
-  useEffect(() => {
-    dispatch(productActions.setProduct(product || null));
-  }, [dispatch, product]);
 
   function handleCloseDetails() {
     const newSearchParams = createSearchParams(appSearchParams);
